refactor(daemon): extract onboarding finalize surface
