+++ crates/daemon/src/lib.rs
@@ -86,6 +86,7 @@ pub mod migrate_cli;
 pub mod migration;
 pub mod next_actions;
 pub mod onboard_cli;
+mod onboard_finalize;
 pub mod onboard_presentation;
 mod provider_credential_policy;
 pub mod provider_presentation;
